Reverted to the old setup
